+++ Modules/TArc/Sources/TArc/Controls/PropertyPanel/Private/ChildCreator.h
@@ -23,7 +23,7 @@ public:
     void RegisterExtension(const std::shared_ptr<ChildCreatorExtension>& extension);
     void UnregisterExtension(const std::shared_ptr<ChildCreatorExtension>& extension);
 
-    Signal<std::shared_ptr<const PropertyNode> /*parent*/, std::shared_ptr<PropertyNode> /*child*/, size_t /*childPosition*/> nodeCreated;
+    Signal<std::shared_ptr<const PropertyNode> /*parent*/, std::shared_ptr<PropertyNode> /*child*/, int32 /*childPosition*/> nodeCreated;
     Signal<std::shared_ptr<PropertyNode> /*child*/> nodeRemoved;
 
 private:
